+++ delete-user.php
@@ -1,27 +1,56 @@
+<!-- This is the delete user which deletes the user record from the database according to user id recieved using $_GET method -->
+
 <?php
+    // Authorization check for private page
     require 'includes/auth.php';
+    // title
     $pageName = 'deleting user data...';
+    // header
     require 'includes/header.php';
 
     try
     {
         $flag = true;
-        $id ='';
-        $id = $_GET['id'];
-        if(!empty($id))
+        
+        $id = '';
+        // checking whether the id parameter is set or not
+        if(isset($_GET['id']) &&  !empty($_GET['id']))
         {
+            $id = $_GET['id'];
+
+            // connecting with the database
             require 'includes/db-conn.php';
-            $sql = 'DELETE FROM users WHERE userId = :id';
+
+            // Checking whether the user id exists or not 
+            $sql = 'SELECT * FROM users WHERE userId = :id';
             $cmd = $db->prepare($sql);
             $cmd->bindParam(':id', $id, PDO::PARAM_INT);
             $cmd->execute();
+            $userPresent = $cmd->fetch();
             $db = null;
-            echo    '<div class="alert alert-success" role="alert">
-                            User record was deleted.   
-                        </div>';
-            //  use this to relocate to the adminstrators list 
-            header("location:administrators.php");
-    
+
+            if($userPresent)
+            {
+                // if user is present the record is deleted and the user list is updated
+                require 'includes/db-conn.php';
+                $sql = 'DELETE FROM users WHERE userId = :id';
+                $cmd = $db->prepare($sql);
+                $cmd->bindParam(':id', $id, PDO::PARAM_INT);
+                $cmd->execute();
+                // Page redirection on successful record deletion
+                echo    '<script>
+                            window.location.replace("administrators.php?update=true");
+                        </script>';
+                $db = null;
+
+            }
+            else
+            {
+                // Page redirection for any errors
+                echo    '<script>
+                            window.location.replace("error.php");
+                        </script>';
+            }
         }
         else
         {
@@ -31,11 +60,10 @@
     }
     catch(Exception $error)
     {
+        // redirection to error page for database errors
         header("location:error.php");
     }
 
-
-?>
-<?php
-  require 'includes/footer.php';
+    // footer file inclusion
+    require 'includes/footer.php';
 ?>
